Removed some unused code, and made some properties internal
