+++ src/Yale/Core/ValueCollection.cs
@@ -7,7 +7,7 @@ using Yale.Core.Interface;
 
 namespace Yale.Core
 {
-    public sealed class ValueCollection : INotifyPropertyChanged
+    internal sealed class ValueCollection : INotifyPropertyChanged
     {
         private readonly IDictionary<string, IValue> _values = new Dictionary<string, IValue>();
 
@@ -42,10 +42,10 @@ namespace Yale.Core
             return success;
         }
 
-        public bool TryGetValue(string key, out IValue value)
+        public bool TryGetValue<T>(string key, out T value)
         {
             var success = _values.TryGetValue(key, out var result);
-            value = result;
+            value = (T)result;
             return success;
         }
